+++ rap/common/conn.py
@@ -45,7 +45,7 @@ class BaseConnection:
         """
         try:
             await asyncio.wait_for(asyncio.shield(self.conn_future), timeout=delay)
-        except (asyncio.TimeoutError, asyncio.CancelledError):
+        except asyncio.TimeoutError:
             pass
 
     async def write(self, data: tuple) -> None:
@@ -141,7 +141,7 @@ class Connection(BaseConnection):
         self.weight: int = weight
         self._ssl_crt_path: Optional[str] = ssl_crt_path
         self.connection_info: str = f"{host}:{port}"
-        self.score: float = 0.0
+        self.score: float = 10.0
 
         self.listen_future: asyncio.Future = done_future()
         self.semaphore: Semaphore = Semaphore(max_conn_inflight or 100)