+++ rap/server/__init__.py
@@ -1,5 +1,5 @@
 from .context import context
 from .core import Server
 from .model import Request, Response
-from .receiver import Channel, Receiver
+from .receiver import Receiver, UserChannel
 from .sender import Sender