+++ tests/test_channel.py
@@ -4,12 +4,13 @@ from typing import Any
 import pytest
 from pytest_mock import MockerFixture
 
-from rap.client import Channel, Client, Response
+from rap.client import Client, Response
+from rap.common.channel import UserChannel
 from rap.common.exceptions import ChannelError, FuncNotFoundError
 from rap.common.utils import Constant
-from rap.server import Channel as ServerChannel
 from rap.server import Response as ServerResponse
 from rap.server import Server
+from rap.server import UserChannel as ServerChannel
 
 pytestmark = pytest.mark.asyncio
 
@@ -19,7 +20,7 @@ class TestChannel:
         msg: str = "hello!"
 
         @rap_client.register()
-        async def async_channel(channel: Channel) -> None:
+        async def async_channel(channel: UserChannel) -> None:
             await channel.write(msg)
             cnt: int = 0
             while await channel.loop(cnt < 3):
@@ -38,7 +39,7 @@ class TestChannel:
 
     async def test_echo_body(self, rap_server: Server, rap_client: Client) -> None:
         @rap_client.register()
-        async def echo_body(channel: Channel) -> None:
+        async def echo_body(channel: UserChannel) -> None:
             msg: str = "hello!"
             cnt: int = 0
             await channel.write(msg)
@@ -49,7 +50,7 @@ class TestChannel:
                 cnt += 1
                 await channel.write(body)
 
-        async def _echo_body(channel: Channel) -> None:
+        async def _echo_body(channel: UserChannel) -> None:
             async for body in channel.iter_body():
                 await channel.write(body)
 
@@ -58,11 +59,11 @@ class TestChannel:
 
     async def test_echo_response(self, rap_server: Server, rap_client: Client) -> None:
         @rap_client.register()
-        async def echo_response(channel: Channel) -> None:
+        async def echo_response(channel: UserChannel) -> None:
             msg: str = "hello!"
             cnt: int = 0
             await channel.write(msg)
-            async for response in channel.iter_response():
+            async for response in channel.iter():
                 # IDE cannot check
                 response: Response = response  # type: ignore
                 assert msg == response.body
@@ -71,8 +72,8 @@ class TestChannel:
                 cnt += 1
                 await channel.write(response.body)
 
-        async def _echo_response(channel: Channel) -> None:
-            async for response in channel.iter_response():
+        async def _echo_response(channel: UserChannel) -> None:
+            async for response in channel.iter():
                 # IDE cannot check
                 response: ServerResponse = response  # type: ignore
                 await channel.write(response.body)
@@ -82,7 +83,7 @@ class TestChannel:
 
     async def test_while_channel_close(self, rap_server: Server, rap_client: Client) -> None:
         @rap_client.register()
-        async def async_channel(channel: Channel) -> None:
+        async def async_channel(channel: UserChannel) -> None:
             await channel.write("hello")
             cnt: int = 0
             while await channel.loop(cnt < 3):
@@ -90,17 +91,12 @@ class TestChannel:
                 print(await channel.read_body())
             return
 
-        async def _async_channel(channel: Channel) -> None:
+        async def _async_channel(channel: UserChannel) -> None:
             while await channel.loop():
                 body: Any = await channel.read_body()
                 if body == "hello":
-                    cnt: int = 0
-                    await channel.close()
+                    break
 
-                    with pytest.raises(ChannelError):
-                        await channel.read_body()
-                    with pytest.raises(ChannelError):
-                        await channel.write(f"hello {cnt}")
                 else:
                     await channel.write("I don't know")
 
@@ -109,11 +105,11 @@ class TestChannel:
             await async_channel()
 
         exec_msg: str = e.value.args[0]
-        assert exec_msg == "recv drop event, close channel"
+        assert exec_msg == "recv channel's drop event, close channel"
 
     async def test_not_found_channel_func(self, rap_server: Server, rap_client: Client) -> None:
         @rap_client.register()
-        async def async_channel(channel: Channel) -> None:
+        async def async_channel(channel: UserChannel) -> None:
             await channel.write("hello")
             cnt: int = 0
             while await channel.loop(cnt < 3):
@@ -137,8 +133,9 @@ class TestChannel:
                 await asyncio.sleep(0.1)
 
         @rap_client.register("test_channel")
-        async def test_client_channel(channel: Channel) -> None:
-            await channel.write("close")
+        async def test_client_channel(channel: UserChannel) -> None:
+            async for response in channel.iter():
+                await channel.write("close")
 
         rap_server.register(test_server_channel, "test_channel")